remove PCI prefix and Id fields
THIS IS A BACKWARDS-INCOMPATIBLE CHANGE

Due to the fact that pcidb used to be a submodule in the ghw library,
all of its publicly-exposed structs used to be prefixed with "PCI". Now
that pcidb is its own library, there's no reason to continue using that
prefix.

Since this is a backwards-incompatible change, I took the opportunity to
remove the old "Id" fields which had been deprecated in the 0.2 release.
No use waiting around to remove those for a 1.0 release when this change
is backwards incompatible anyway.

Issue #9