do not error when $HOME not set
In calculating the default cache path, we were using the
`os.UserHomeDir()` function to get the user's home directory. This
function differs from the behaviour of mitchellh/go-homedir in that it
errors out on Linux if `$HOME` is not set.

This caused issues in certain environments and so this PR simply returns
an empty string when `$HOME` is unset and now returns an error only
during pci.ids DB file discovery if search paths yields an empty slice
AND cache path is empty AND network fetch is enabled.

Issue #38

Signed-off-by: Jay Pipes <jaypipes@gmail.com>